mv function pkg from fn-harness to java-core
