+++ sdks/java/core/src/main/java/org/apache/beam/sdk/function/ThrowingBiConsumer.java
@@ -15,17 +15,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.fn.function;
+package org.apache.beam.sdk.function;
 
-import java.util.function.Consumer;
+import java.util.function.BiConsumer;
 
 /**
- * A {@link Consumer} which can throw {@link Exception}s.
+ * A {@link BiConsumer} which can throw {@link Exception}s.
  *
  * <p>Used to expand the allowed set of method references to be used by Java 8 functional
  * interfaces.
  */
 @FunctionalInterface
-public interface ThrowingConsumer<T> {
-  void accept(T t) throws Exception;
+public interface ThrowingBiConsumer<T1, T2> {
+  void accept(T1 t1, T2 t2) throws Exception;
 }
