+++ sdks/java/core/src/main/java/org/apache/beam/sdk/function/ThrowingBiFunction.java
@@ -15,6 +15,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.sdk.function;
 
-/** Java 8 functional interface extensions. */
-package org.apache.beam.sdk.fn.function;
+import java.util.function.BiFunction;
+
+/**
+ * A {@link BiFunction} which can throw {@link Exception}s.
+ *
+ * <p>Used to expand the allowed set of method references to be used by Java 8 functional
+ * interfaces.
+ */
+@FunctionalInterface
+public interface ThrowingBiFunction<T1, T2, T3> {
+  T3 apply(T1 t1, T2 t2) throws Exception;
+}
