+++ sdks/java/core/src/main/java/org/apache/beam/sdk/function/ThrowingFunction.java
@@ -15,17 +15,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.fn.function;
+package org.apache.beam.sdk.function;
 
-import java.util.function.BiFunction;
+import java.util.function.Function;
 
 /**
- * A {@link BiFunction} which can throw {@link Exception}s.
+ * A {@link Function} which can throw {@link Exception}s.
  *
  * <p>Used to expand the allowed set of method references to be used by Java 8 functional
  * interfaces.
  */
 @FunctionalInterface
-public interface ThrowingBiFunction<T1, T2, T3> {
-  T3 apply(T1 t1, T2 t2) throws Exception;
+public interface ThrowingFunction<T1, T2> {
+  T2 apply(T1 value) throws Exception;
 }
