+++ sdks/java/core/src/main/java/org/apache/beam/sdk/function/ThrowingRunnable.java
@@ -15,17 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.fn.function;
-
-import java.util.function.Function;
+package org.apache.beam.sdk.function;
 
 /**
- * A {@link Function} which can throw {@link Exception}s.
+ * A {@link Runnable} which can throw {@link Exception}s.
  *
  * <p>Used to expand the allowed set of method references to be used by Java 8 functional
  * interfaces.
  */
 @FunctionalInterface
-public interface ThrowingFunction<T1, T2> {
-  T2 apply(T1 value) throws Exception;
+public interface ThrowingRunnable {
+  void run() throws Exception;
 }
